backend/requirements.txt
google-cloud-vision==3.9.0
Flask==3.1.0
gunicorn==23.0.0
supabase==2.10.0
python-dotenv==1.0.1
google-cloud-aiplatform==1.71.1
spacy==3.8.3
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
Flask-Cors==5.0.0
numpy==2.0.0
vertexai==1.71.1